Added a script for proxy settings
This script enable customer to setup to setup proxy settings for the Jupyter notebooks and the SageMaker notebook instance. In order for the Jupyter notebooks to be able to pick up the proxy settings from the Lifecycle Configuration, we need to specify these settings in the ipython notebook server environment. To configure the ipython notebook server profile, we will create a simple python script 00-startup.py under /home/ec2-user/.ipython/profile_default/startup path (the 00 in the script name is because scripts under this path are run in lexicographical order, and we want this to be run first). 

Additionally, the default terminal shell that is spawned whenever you open a terminal in Jupyter is a non-logging shell, i.e., it is an interactive shell. Therefore, for the proxy settings to be apply to the shell environment as well via the Lifecycle Configuration, we need to change the default interactive shell to a logging shell like bash and put the proxy setting in /home/ec2-user/.profile.