Updated on-start.sh script
I updated the proxy script to extract the proxy server URL to a variable SERVER, so it's easier to see where you need to replace the sever with your own proxy server and only need to do it once. Additionally, I have added an explanation in the script on why we are excluding S3.