Add script for setting env variable and configuring Git
**Description**
Add script for setting env variable and configuring Git

**Motivation**
Add a way to set env varible in lifecycle config so it could be used in all Jupyter envs. Add a way to set Git config. For now user is default to ec2-user and email is default to host ip.

**Testing Done**
* Verified that env variable is set
* Verified that username and email are set in Git config